+++ ConsoleZapp/CzappTester/Tests/PrintSafetyTests.cs
@@ -1,8 +1,8 @@
 using ConsoleZapp;
 
-namespace CzappTester.Tests
+namespace CzappTester.Tests.Basic
 {
-    public static class Test04_PrintSafety
+    public static class PrintSafetyTests
     {
         public static void DoubleFormatFix()
         {