+++ ConsoleZapp/NugetLib/Controls/BarColor.cs
@@ -1,0 +1,11 @@
+namespace ConsoleZapp
+{
+    public enum BarColor
+    {
+        White,
+        Green,
+        Yellow,
+        LightBlue,
+        Red
+    }
+}