+++ ConsoleZapp/NugetLib/Controls/Base/IControl.cs
@@ -1,0 +1,8 @@
+namespace ConsoleZapp
+{
+    public interface IControl
+    {
+        string Render();
+        void Print();
+    }
+}